Squashed 'data_compression/' changes from be34553..422b6e7
422b6e7 Merge branch 'master' of gitenterprise.xilinx.com:FaaSApps/xf_compression into next
9acef90 ZLIB/GZip: 4gb limitation issue (#852)
4a9bca7 Merge branch 'master' of gitenterprise.xilinx.com:FaaSApps/xf_compression into next
be29e13 GZip/Zlib decompression standard header checks (#847)
be7c070 Jenkins standard fix (#854)
5f92db7 Merge branch 'master' of gitenterprise.xilinx.com:FaaSApps/xf_compression into next
22a47da Zlib/GZip bandwidth benchmark demo updates (#855)
6def687 Merge branch 'master' of gitenterprise.xilinx.com:FaaSApps/xf_compression into next
e4b2dc9 Updating xf_compresson with platform_blacklist support in library.json (#853)
1c8d0db Updating xf_compression with latest autogen script (#850)
3684602 fixed host issue (#851)
ec5ead8 sp tag bank assignments (#846)
d4fdb6c updated based on francis feedback (results reporting mbps --> MBps) (#848)
82226aa fixed list issue and compressed file can be sent as input (#845)
9221961 Merge branch 'master' of gitenterprise.xilinx.com:FaaSApps/xf_compression into next
4fe2e10 Putting vck in platform_blacklist (#835)
335fc7a Merge branch 'master' of gitenterprise.xilinx.com:FaaSApps/xf_compression into next
965feed updated sequential host to enqueue all data in one call (#840)
e24f89d BRAM is used for lz4 litStream depth (#841)
b563069 zlib_config to zlib_specs in L1/include (#812)
2251c0f removed duplicate defines in lz4 p2p compression (#838)
6304585 lz4 header processing as part of multibyte decompress kernel (#827)
af73024 Update Jenkinsfile
4a56f30 Update Jenkinsfile
935c653 Update Jenkinsfile
058b16a Merge branch 'master' of gitenterprise.xilinx.com:FaaSApps/xf_compression into next
e6e66e0 snappy multi byte fmax improvements (#831)

git-subtree-dir: data_compression
git-subtree-split: 422b6e7f14c49d688936a074ec22676872cbf839

+++ L2/tests/src/zlib.hpp
@@ -28,7 +28,7 @@
 #include <string>
 #include <fstream>
 #include "xcl2.hpp"
-#include "zlib_config.hpp"
+#include "zlib_specs.hpp"
 #include <thread>
 
 #pragma once
@@ -75,13 +75,13 @@ void error_message(const std::string& val);
 
 int validate(std::string& inFile_name, std::string& outFile_name);
 
-uint32_t get_file_size(std::ifstream& file);
+uint64_t get_file_size(std::ifstream& file);
 
 class xil_zlib {
    public:
     int init(const std::string& binaryFile, uint8_t flow, uint8_t d_type);
     int release();
-    uint32_t compress(uint8_t* in, uint8_t* out, uint32_t actual_size, uint32_t host_buffer_size);
+    uint32_t compress(uint8_t* in, uint8_t* out, uint64_t actual_size, uint32_t host_buffer_size);
     uint32_t decompress(uint8_t* in, uint8_t* out, uint32_t actual_size, int cu_run, bool enable_p2p = 0);
     uint32_t compress_file(std::string& inFile_name, std::string& outFile_name, uint64_t input_size);
     uint32_t decompress_file(
